# Speed up Colimit_Flattening