On my machine, the time goes from 3.6s to 1.3s.  There are only two slow lines, both marked with TODO, but they are the tricky kind:  a Defined (0.4s), and the closing of the Section (0.2s).  Normally I would just remove the section, but if I simply remove it and leave the Context command, then the file becomes slower overall.  It might work to copy the Context to every definition, but I didn't think that was worth it for 0.2s.