@jdchristensen 2023-10-20
The second commit is only indentation, so look at the other two commits separately.