@Alizter 2023-10-20
I wonder what we are spending most of the time on?