@jdchristensen 2023-10-20
> I wonder what we are spending most of the time on?

One of the proofs is still quite long, with lots of path algebra.  To make it faster, one would probably need to abstract away some of the situation into one or more helper lemmas.  But these simpler changes at least make a big dent in the time.