@JasonGross 2023-10-21
> but if I simply remove it and leave the Context command, then the file becomes slower overall

This might be because `Context` outside a `Section` actually means `Axiom`...