@jdchristensen 2023-10-21
> This might be because `Context` outside a `Section` actually means `Axiom`...

@JasonGross Because of that, removing the `Section` does not have the semantics we want, but it's still very surprising to me that it makes the file process more slowly.  Do you have some intuition or insight about why having a few axioms would cause a file to be slower than have a `Section` with `Context`?