@Alizter 2023-10-21
@jdchristensen It's been a while since I thought about this kind of thing, but my intuition is that axioms and qeds are generally slower for the Coq kernel to handle since it needs to backtrack on its reduction strategy.

When you close a section, the terms in the section go through a process known internally as "cooking" which makes sure that the correct context variables are abstracted as lambdas in order to make a closed term. Notably the difference between the non-section term is the fact that the axioms are now variables.

So I would guess that when working with an axiom, it is somehow treated differently than working with a variable hence the difference in speed.

That's what I imagine is going on, but I could be totally wrong. @SkySkimmer might be able to correct me or offer an insight.