@JasonGross 2023-10-21
> > This might be because `Context` outside a `Section` actually means `Axiom`...
> 
> @JasonGross Because of that, removing the `Section` does not have the semantics we want, but it's still very surprising to me that it makes the file process more slowly. Do you have some intuition or insight about why having a few axioms would cause a file to be slower than have a `Section` with `Context`?

Two guesses:
1. Context variables have a single universe assigned to them for the whole section, while axioms can be universe polymorphic.  So maybe removing the section added a bunch of universe variables everywhere
2. If any of the context variables got introduced as typeclasses maybe they persisted past the close of the original section and slowed down later typeclass search