@JasonGross 2023-10-21
Also, @Alizter I don't see how your description could be accurate.  The closest thing to that I can imagine happening is that maybe some later tactic or unification heuristic is extremely sensitive to how many arguments some application node has, and so having your constant be applied to, e.g., only 5 things instead of 7 because the first two became axioms rather than arguments made some heuristic take a wrong path.