@jdchristensen 2023-10-21
@JasonGross I should have pointed out that the entire file was wrapped in that section, so there is no later material that could get slower when the section is removed.  That's what's puzzling to me.  Since the `End` of the section took a bit of time, it seemed to me like removing the section should only help things...